refactor(server): upload config
